+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+2.0.0 (2020-08-02)
+------------------
+* Method names are now conform the ``pyqtgraph`` naming style. I.e.
+  ``setData()`` vs. ``set_data()``, etc.
+* The default values of ``pyqtgraph`` are no longer being overwritten.
+* Added class ``LegendSelect``
+
 1.0.0 (2020-07-30)
 ------------------
 * First release on PyPI